requirements.txt
# django~=2.2
pillow~=7.1.2
django-import-export~=2.1.0